fix: resolve slash command index mismatch between dropdown and keyboard handler
Extract shared filterAndSortSlashCommands() utility to ensure both InputArea
(visual dropdown) and useInputKeyDown (Enter/Tab handler) produce identically
ordered command lists. Previously, InputArea sorted by relevance score while
useInputKeyDown kept original array order, causing the wrong command to be
inserted when pressing Enter/Tab.

Also fixes SessionList group auto-expansion to use fuzzyMatch() consistently
instead of substring .includes(), matching the main sessionCategories filter.
